removed un used custom resolvers and moved resolvers into own folders
