+++ resolver-functions/Ziplink/ZipLinkQuery.test.ts
@@ -1,8 +1,8 @@
 import { ExecuteMeshFn } from "@graphql-mesh/runtime";
-import { getExampleQuery } from "./utils/ExampleQueryFiles";
-import { getMeshInstance } from "./utils/MeshInstance";
+import { getExampleQuery } from "../../tests/utils/ExampleQueryFiles";
+import { getMeshInstance } from "../../tests/utils/MeshInstance";
 
-import * as httpUtils from "../utils/httpUtils";
+import * as httpUtils from "../../utils/httpUtils";
 jest.mock("../utils/httpUtils");
 
 beforeEach(() => {
