+++ utils/constants.ts
@@ -1,0 +1,5 @@
+/**
+ * Arbitrary very large number used temporarily during Rails read phase to force Rails not to
+ * paginate our fake "Workflows Service" call.
+ */
+export const TEN_MILLION = 10_000_000;
